refactor: move AnyTransactionReceipt to dedicated transaction.rs file
